Move ChapterAdapter into chapter package
